refactor(data): move searchLocationsModelFactory to helpers
