+++ tests/data/remoteSearchLocations.spec.ts
@@ -1,12 +1,10 @@
-import { faker } from '@faker-js/faker';
 import { HttpStatusCode } from '~/data/http';
 import { Filter, Ordination } from '~/domain/enums';
 import { NoAccessError, UnexpectedError } from '~/data/errors';
 import { RemoteSearchLocations } from '~/data/useCases';
-import SearchLocationModel from '../../src/domain/models/searchLocationModel';
 import { makeUrl } from './helpers/testFactories';
 import { HttpClientSpy } from './http/httpClientSpy';
-import businessHoursModelStub from './stubs/businessHoursModelStub';
+import searchLocationsModelFactory from './helpers/searchLocationsModelFactory';
 
 describe('Data: RemoteSearchLocations', () => {
   test('should search with httpPostClient calling correct url with page param', () => {
@@ -109,38 +107,4 @@ const makeSut = (url = makeUrl()) => {
   const sut = new RemoteSearchLocations(url, httpClient);
 
   return { sut, httpClient };
-};
-
-const searchLocationsModelFactory = (): SearchLocationModel[] => {
-  return [
-    {
-      about: faker.lorem.paragraph(),
-      address: faker.address.streetAddress(),
-      businessHours: businessHoursModelStub(),
-      comments: makeComments(),
-      contact: faker.phone.number(),
-      images: makeImages(),
-      isSaved: true,
-      name: faker.company.name(),
-      rating: 5,
-    },
-  ];
-};
-
-const makeComments = (quantity = 5): string[] => {
-  const comments = [];
-  for (let index = 0; index < quantity; index++) {
-    comments.push(faker.lorem.paragraph());
-  }
-
-  return comments;
-};
-
-const makeImages = (quantity = 5): string[] => {
-  const images = [];
-  for (let index = 0; index < quantity; index++) {
-    images.push(faker.image.imageUrl());
-  }
-
-  return images;
 };
